@@ -1,4 +1,4 @@
-# Copyright 2021 - 2023 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
+# Copyright 2021 - 2024 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 # for the German Human Genome-Phenome Archive (GHGA)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,7 +18,6 @@
 
 from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Optional
 
 from ghga_service_commons.utils.context import asyncnullcontext
 from hexkit.providers.akafka.provider import KafkaEventSubscriber
@@ -49,7 +48,7 @@ async def prepare_core(*, config: Config) -> AsyncGenerator[NotifierPort, None]:
 
 
 def prepare_core_with_override(
-    *, config: Config, notifier_override: Optional[NotifierPort] = None
+    *, config: Config, notifier_override: NotifierPort | None = None
 ):
     """Resolve the notifier context manager based on config and override (if any)."""
     return (
@@ -63,7 +62,7 @@ def prepare_core_with_override(
 async def prepare_event_subscriber(
     *,
     config: Config,
-    notifier_override: Optional[NotifierPort] = None,
+    notifier_override: NotifierPort | None = None,
 ) -> AsyncGenerator[KafkaEventSubscriber, None]:
     """Construct and initialize an event subscriber with all its dependencies.
     By default, the core dependencies are automatically prepared but you can also
